+++ src/mouse_core/models/heads/__init__.py
@@ -1,4 +1,4 @@
-from mouse_core.models.heads.base import BaseHead, BaseHeadWithTarget, HeadSpec
+from mouse_core.models.heads.base import BaseHead, HeadSpec
 from mouse_core.models.heads.swiglu import SwiGLUHead
 from mouse_core.models.heads.layerwise_dqn import LayerwiseDiscreteActionValueHead
 from mouse_core.models.heads.discrete_action import DiscreteActionHead
@@ -6,7 +6,6 @@ from mouse_core.models.heads.dqn import DiscreteActionValueHead
 
 __all__ = [
     "BaseHead",
-    "BaseHeadWithTarget",
     "HeadSpec",
     "SwiGLUHead",
     "DiscreteActionHead",